refactor: cleanup, removed default exports
