+++ apps/strapi/src/components/footer/footer-cta-badge.json
