+++ apps/strapi/src/components/footer/footer-cta-card.json
