+++ apps/strapi/src/components/footer/footer-main.json
