+++ apps/ui/src/app/[locale]/dev/components-overview/page.tsx
@@ -2,8 +2,8 @@ import { uniq } from "lodash"
 import type { Locale } from "next-intl"
 import { setRequestLocale } from "next-intl/server"
 
-import ComponentsList from "@/app/[locale]/dev/components-overview/components/ComponentsList"
-import Typography from "@/components/typography"
+import { ComponentsList } from "@/app/[locale]/dev/components-overview/components/ComponentsList"
+import { Typography } from "@/components/typography"
 import { logNonBlockingError } from "@/lib/logging"
 import { PublicStrapiClient } from "@/lib/strapi-api"
