+++ apps/ui/src/app/[locale]/dev/layout.tsx
@@ -1,6 +1,6 @@
 import { notFound } from "next/navigation"
 
-import DevNavbar from "@/app/[locale]/dev/components/DevNavbar"
+import { DevNavbar } from "@/app/[locale]/dev/components/DevNavbar"
 import { Container } from "@/components/elementary/Container"
 import { isProduction } from "@/lib/general-helpers"
