+++ apps/ui/src/app/[locale]/dynamic/[[...rest]]/page.tsx
@@ -2,7 +2,7 @@ import { ROOT_PAGE_PATH } from "@repo/shared-data"
 import type { Locale } from "next-intl"
 import { use } from "react"
 
-import StrapiPageView from "@/components/layouts/StrapiPageView"
+import { StrapiPageView } from "@/components/layouts/StrapiPageView"
 import { getMetadataFromStrapi } from "@/lib/metadata"
 
 // Force dynamic rendering (SSR) for this route
