+++ apps/ui/src/components/page-builder/components/navigation/navbar/MobileNavbar.tsx
@@ -2,10 +2,9 @@ import type { Nullable } from "@repo/shared-data"
 import type { Data } from "@repo/strapi-types"
 
 interface MobileNavbarProps {
-  readonly logoImage: Nullable<Data.Component<"utilities.link-image">>
   readonly navItems: Nullable<Data.Component<"navbar.nav-item">[]>
 }
 
-export function MobileNavbar({ logoImage, navItems }: MobileNavbarProps) {
+export function MobileNavbar({ navItems }: MobileNavbarProps) {
   return <div className="flex h-20 items-center lg:hidden" />
 }
