+++ apps/ui/src/components/page-builder/components/plans/StrapiPlanComparisonTable.tsx
@@ -311,7 +311,3 @@ export function StrapiPlanComparisonTable({
     </Container>
   )
 }
-
-StrapiPlanComparisonTable.displayName = "StrapiPlanComparisonTable"
-
-export default StrapiPlanComparisonTable
