+++ apps/ui/src/components/page-builder/components/plans/StrapiPlanPricingCards/PricingCheckoutModal.tsx
@@ -238,7 +238,3 @@ export function PricingCheckoutModal({ card }: PricingCheckoutModalProps) {
     </Dialog>
   )
 }
-
-PricingCheckoutModal.displayName = "PricingCheckoutModal"
-
-export default PricingCheckoutModal
