+++ apps/ui/src/components/page-builder/components/plans/StrapiPlanPricingCards/StrapiPlanPricingCards.tsx
@@ -114,7 +114,3 @@ export function StrapiPlanPricingCards({
     </Container>
   )
 }
-
-StrapiPlanPricingCards.displayName = "StrapiPlanPricingCards"
-
-export default StrapiPlanPricingCards
