+++ apps/ui/src/components/page-builder/components/sections/StrapiHowItWorks.tsx
@@ -72,7 +72,3 @@ export function StrapiHowItWorks({
     </section>
   )
 }
-
-StrapiHowItWorks.displayName = "StrapiHowItWorks"
-
-export default StrapiHowItWorks
