+++ apps/ui/src/components/page-builder/components/sections/StrapiIntegrationsSection.tsx
@@ -139,7 +139,3 @@ export function StrapiIntegrationsSection({
     </section>
   )
 }
-
-StrapiIntegrationsSection.displayName = "StrapiIntegrationsSection"
-
-export default StrapiIntegrationsSection
