+++ apps/ui/src/components/page-builder/components/utilities/StrapiLinkText.tsx
@@ -72,7 +72,3 @@ export function StrapiLinkText({
     </a>
   )
 }
-
-StrapiLinkText.displayName = "StrapiLinkText"
-
-export default StrapiLinkText
