+++ apps/ui/src/components/page-builder/index.tsx
@@ -1,14 +1,14 @@
 import type { UID } from "@repo/strapi-types"
 
-import StrapiNewsletterForm from "@/components/page-builder/components/forms/StrapiNewsletterForm"
-import StrapiNavbarComponent from "@/components/page-builder/components/navigation/StrapiNavbarComponent"
-import StrapiPlanComparisonTable from "@/components/page-builder/components/plans/StrapiPlanComparisonTable"
-import StrapiPlanPricingCards from "@/components/page-builder/components/plans/StrapiPlanPricingCards/StrapiPlanPricingCards"
-import StrapiBannerSlice from "@/components/page-builder/components/sections/StrapiBannerSlice"
-import StrapiFaqSection from "@/components/page-builder/components/sections/StrapiFaqSection"
-import StrapiHowItWorks from "@/components/page-builder/components/sections/StrapiHowItWorks"
-import StrapiIntegrationsSection from "@/components/page-builder/components/sections/StrapiIntegrationsSection"
-import StrapiUserStoriesSection from "@/components/page-builder/components/sections/StrapiUserStoriesSection"
+import { StrapiNewsletterForm } from "@/components/page-builder/components/forms/StrapiNewsletterForm"
+import { StrapiNavbar } from "@/components/page-builder/components/navigation/navbar/StrapiNavbar"
+import { StrapiPlanComparisonTable } from "@/components/page-builder/components/plans/StrapiPlanComparisonTable"
+import { StrapiPlanPricingCards } from "@/components/page-builder/components/plans/StrapiPlanPricingCards/StrapiPlanPricingCards"
+import { StrapiBannerSlice } from "@/components/page-builder/components/sections/StrapiBannerSlice"
+import { StrapiFaqSection } from "@/components/page-builder/components/sections/StrapiFaqSection"
+import { StrapiHowItWorks } from "@/components/page-builder/components/sections/StrapiHowItWorks"
+import { StrapiIntegrationsSection } from "@/components/page-builder/components/sections/StrapiIntegrationsSection"
+import { StrapiUserStoriesSection } from "@/components/page-builder/components/sections/StrapiUserStoriesSection"
 import { StrapiFooterCta } from "@/components/page-builder/single-types/footer/StrapiFooterCta"
 import { StrapiFooterMain } from "@/components/page-builder/single-types/footer/StrapiFooterMain"
 
@@ -37,9 +37,9 @@ export const ContentComponents: Partial<
   "sections.user-stories-section": StrapiUserStoriesSection,
 
   // Footer
-  "sections.footer-main": StrapiFooterMain,
-  "sections.footer-cta": StrapiFooterCta,
+  "footer.footer-main": StrapiFooterMain,
+  "footer.footer-cta": StrapiFooterCta,
 
   // Navigation
-  "navigation.navbar": StrapiNavbarComponent,
+  "navigation.navbar": StrapiNavbar,
 }
