+++ apps/ui/src/components/typography/index.tsx
@@ -116,7 +116,3 @@ export function Typography({
     </Tag>
   )
 }
-
-Typography.displayName = "Typography"
-
-export default Typography
